refactor: extract session ownership and align lifecycle docs
- Move managed/explicit cleanup registries into session-ownership module
- Document explicit open cleanup on quit and reload across contracts/playbook
