+++ extensions/agent-browser/index.ts
@@ -37,14 +37,24 @@ import {
 } from "./lib/runtime.js";
 import { isRecord } from "./lib/parsing.js";
 import { buildPromptPolicy, getLatestUserPrompt, shouldAppendBrowserSystemPrompt } from "./lib/prompt-policy.js";
-import { isCloseCommand, isOpenNavigationCommand } from "./lib/command-taxonomy.js";
+import { isCloseCommand } from "./lib/command-taxonomy.js";
 import { cleanupSecureTempArtifacts } from "./lib/temp.js";
 import {
 	AGENT_BROWSER_PARAMS,
 	type CompiledAgentBrowserElectron,
 } from "./lib/input-modes.js";
 import { parseAllowedDomainsPolicyFromArgs, type AllowedDomainsPolicy } from "./lib/navigation-policy.js";
-import { closeManagedSession, getSessionContextKey, runAgentBrowserTool, type AgentBrowserToolResult, type BrowserRunState, type TraceOwner } from "./lib/orchestration/browser-run/index.js";
+import { getSessionContextKey, runAgentBrowserTool, type BrowserRunState, type TraceOwner } from "./lib/orchestration/browser-run/index.js";
+import {
+	closeOwnedManagedSessions,
+	closeOwnedManagedSessionsExcept,
+	syncExplicitCleanupSessionsFromResult,
+	syncOwnedManagedSessionsFromResult,
+	trackOwnedManagedSession,
+	untrackOwnedManagedSession,
+	untrackOwnedManagedSessionFromBranchClose,
+	type OwnedManagedSession,
+} from "./lib/orchestration/browser-run/session-ownership.js";
 import { findElectronLaunchRecordForSession, getActiveElectronRecords } from "./lib/orchestration/browser-run/session-state.js";
 import { parseBatchStdinJsonArray } from "./lib/orchestration/batch-stdin.js";
 import {
@@ -83,13 +93,6 @@ function isBashToolCallEvent(event: unknown): event is BashToolCallLike {
 	if (!isRecord(event) || event.toolName !== "bash" || !isRecord(event.input)) return false;
 	return typeof event.input.command === "string";
 }
-
-type OwnedManagedSession = {
-	branchOwned: boolean;
-	cwd: string;
-	namespace?: string;
-	sessionName: string;
-};
 
 // Event ranks are local to the branch being restored. Keep them out of owned-resource
 // state so branch switches never compare unrelated branch histories.
@@ -189,96 +192,6 @@ function restoreAllowedDomainsBySessionFromBranch(branch: unknown[]): Map<string
 		if (policy && sessionKey) restoredPolicies.set(sessionKey, policy);
 	}
 	return restoredPolicies;
-}
-
-function trackOwnedManagedSession(
-	sessions: Map<string, OwnedManagedSession>,
-	sessionName: string | undefined,
-	cwd: string,
-	options: { branchOwned?: boolean; namespace?: string } = {},
-): void {
-	if (!sessionName) return;
-	const key = getSessionContextKey(sessionName, options.namespace) ?? sessionName;
-	const existing = sessions.get(key);
-	const branchOwned = existing && !existing.branchOwned ? false : options.branchOwned === true;
-	sessions.set(key, { branchOwned, cwd, namespace: options.namespace, sessionName });
-}
-
-function untrackOwnedManagedSession(sessions: Map<string, OwnedManagedSession>, sessionName: string | undefined, namespace?: string): void {
-	if (!sessionName) return;
-	if (sessionName.includes("\u0000")) sessions.delete(sessionName);
-	else sessions.delete(getSessionContextKey(sessionName, namespace) ?? sessionName);
-}
-
-function untrackOwnedManagedSessionFromBranchClose(
-	sessions: Map<string, OwnedManagedSession>,
-	sessionName: string | undefined,
-	activeBranchRank: number | undefined,
-	closeBranchRank: number | undefined,
-): void {
-	if (!sessionName || closeBranchRank === undefined) return;
-	const ownedSession = sessions.get(sessionName);
-	if (!ownedSession?.branchOwned) return;
-	if (activeBranchRank !== undefined && closeBranchRank <= activeBranchRank) return;
-	sessions.delete(sessionName);
-}
-
-function syncOwnedManagedSessionsFromResult(
-	sessions: Map<string, OwnedManagedSession>,
-	explicitCleanupSessions: Map<string, OwnedManagedSession>,
-	result: AgentBrowserToolResult,
-	cwd: string,
-): void {
-	const details = isRecord(result.details) ? result.details : undefined;
-	const outcome = isRecord(details?.managedSessionOutcome) ? details.managedSessionOutcome : undefined;
-	if (!outcome) return;
-	const namespace = isRecord(details) && typeof details.namespace === "string" ? details.namespace : undefined;
-	const succeeded = outcome.succeeded === true;
-	const status = typeof outcome.status === "string" ? outcome.status : undefined;
-	const currentSessionName = typeof outcome.currentSessionName === "string" ? outcome.currentSessionName : undefined;
-	const attemptedSessionName = typeof outcome.attemptedSessionName === "string" ? outcome.attemptedSessionName : undefined;
-	if (outcome.activeAfter === true && (status === "created" || status === "replaced" || status === "unchanged")) {
-		trackOwnedManagedSession(sessions, currentSessionName, cwd, { namespace });
-		// Managed ownership wins; keep the two registries disjoint.
-		untrackOwnedManagedSession(explicitCleanupSessions, currentSessionName, namespace);
-	}
-	if (succeeded && status === "closed") {
-		untrackOwnedManagedSession(sessions, attemptedSessionName ?? currentSessionName, namespace);
-	}
-}
-
-// Process-owned cleanup for caller explicit --session opens. Not on the managed serialize path.
-// Closed on every extension shutdown (quit/reload) because reload starts a new instance.
-function syncExplicitCleanupSessionsFromResult(
-	sessions: Map<string, OwnedManagedSession>,
-	managedSessions: Map<string, OwnedManagedSession>,
-	result: AgentBrowserToolResult,
-	cwd: string,
-): void {
-	const details = isRecord(result.details) ? result.details : undefined;
-	const namespace = isRecord(details) && typeof details.namespace === "string" ? details.namespace : undefined;
-	const sessionName = typeof details?.sessionName === "string" ? details.sessionName : undefined;
-	const command = typeof details?.command === "string" ? details.command : extractCommandTokens(getToolResultArgs(details ?? {})).at(0);
-	const toolSucceeded = result.isError !== true && details?.resultCategory !== "failure";
-	const usedImplicitSession = details?.usedImplicitSession === true;
-	const outcome = isRecord(details?.managedSessionOutcome) ? details.managedSessionOutcome : undefined;
-	if (outcome?.succeeded === true && outcome.status === "closed") {
-		const closedName = typeof outcome.attemptedSessionName === "string"
-			? outcome.attemptedSessionName
-			: typeof outcome.currentSessionName === "string"
-				? outcome.currentSessionName
-				: undefined;
-		untrackOwnedManagedSession(sessions, closedName, namespace);
-	}
-	if (toolSucceeded && sessionName && isCloseCommand(command)) {
-		untrackOwnedManagedSession(sessions, sessionName, namespace);
-		return;
-	}
-	if (!toolSucceeded || !sessionName || usedImplicitSession || !isOpenNavigationCommand(command)) return;
-	const key = getSessionContextKey(sessionName, namespace) ?? sessionName;
-	// Never dual-own a managed session name (implicit or fresh).
-	if (managedSessions.has(key) || outcome) return;
-	trackOwnedManagedSession(sessions, sessionName, cwd, { namespace });
 }
 
 function getTouchedElectronLaunchIds(sessionName: string | undefined, records: Map<string, ElectronLaunchRecord>): Set<string> | undefined {
@@ -519,19 +432,6 @@ function syncElectronCleanupManagedSessions(sessions: Map<string, OwnedManagedSe
 	for (const sessionName of getCleanupResultsClosedManagedSessionNames(cleanupResults)) {
 		untrackOwnedManagedSession(sessions, sessionName);
 	}
-}
-
-async function closeOwnedManagedSessionsExcept(sessions: Map<string, OwnedManagedSession>, keepSessionName: string | undefined, timeoutMs: number, keepNamespace?: string): Promise<void> {
-	const keepKey = getSessionContextKey(keepSessionName, keepNamespace);
-	for (const [key, owner] of [...sessions]) {
-		if (key === keepKey) continue;
-		const error = await closeManagedSession({ cwd: owner.cwd, namespace: owner.namespace, sessionName: owner.sessionName, timeoutMs });
-		if (!error) sessions.delete(key);
-	}
-}
-
-async function closeOwnedManagedSessions(sessions: Map<string, OwnedManagedSession>, timeoutMs: number): Promise<void> {
-	await closeOwnedManagedSessionsExcept(sessions, undefined, timeoutMs);
 }
 
 function getOffBranchOwnedElectronLaunchRecords(ownedRecords: Map<string, ElectronLaunchRecord>, branchRecords: Map<string, ElectronLaunchRecord>): Map<string, ElectronLaunchRecord> {
