+++ extensions/agent-browser/lib/playbook.ts
@@ -45,7 +45,7 @@ export const SHARED_BROWSER_PLAYBOOK_GUIDELINES = [
 	"Do not assume Playwright selector dialects such as text=Close or button:has-text('Close') are supported wrapper syntax unless current upstream agent-browser behavior has been verified.",
 	"For authenticated or user-specific content explicitly requested by the user, such as feeds, inboxes, account pages, or private dashboards, use a real profile only when the user/config asks for it or profiles have been inspected; do not assume --profile Default exists on every machine. Do not use a real profile for public pages just because they are dashboards. Treat visible page content from real profiles as model-visible transcript data; use --auto-connect only if profile-based reuse is unavailable or the task is specifically about attaching to a running debug-enabled browser. If profile/user-data-dir resolution fails, stop retrying opens, run profiles and/or doctor through agent_browser, then report what the user needs to configure.",
 	"Do not invent fixed explicit session names for routine tasks. Use the implicit session unless you truly need multiple isolated browser sessions in the same conversation.",
-	"Always close browser work with close/quit/exit when finished. The wrapper tracks managed and explicit sessions opened through agent_browser and best-effort closes them on Pi quit; idle timeout (default 15 minutes via AGENT_BROWSER_IDLE_TIMEOUT_MS / PI_AGENT_BROWSER_IMPLICIT_SESSION_IDLE_TIMEOUT_MS) is the crash backstop. pi-agent-browser-doctor warns about live /tmp/piab*.pid daemons.",
+	"Always close browser work with close/quit/exit when finished. The wrapper tracks managed sessions and caller explicit open/goto/navigate --session launches; it best-effort closes managed sessions on Pi quit and closes tracked explicit opens on every extension shutdown (quit or /reload, because reload starts a new instance). Idle timeout (default 15 minutes via AGENT_BROWSER_IDLE_TIMEOUT_MS / PI_AGENT_BROWSER_IMPLICIT_SESSION_IDLE_TIMEOUT_MS) is the crash backstop. pi-agent-browser-doctor warns about live /tmp/piab*.pid sidecars.",
 	`When using launch-scoped flags (${LAUNCH_SCOPED_FLAG_LABEL}), put them on the first command for that session. If you intentionally use an explicit --session, keep using that same explicit session for follow-ups.`,
 	`If you already used the implicit session and now need launch-scoped flags (${LAUNCH_SCOPED_FLAG_LABEL}), retry with top-level sessionMode set to fresh or pass an explicit --session for the new launch; never pass --session-mode inside args. After a successful unnamed fresh launch, later auto calls follow that new session.`,
 	"For WebGPU pages, use args [\"--webgpu\", \"open\", \"<url>\"] on a fresh local browser launch; use doctor --webgpu (or --headed on Linux/Windows capture paths) to prove rendering before trusting a non-black screenshot. WebGPU cannot be combined with --cdp, --auto-connect, or provider launches unless --webgpu false overrides an enabled config/environment default.",
